tech(api): delete dead code and move some files in evaluation BC --
