+++ api/src/evaluation/infrastructure/adapters/solution-adapter.js
