+++ api/src/evaluation/infrastructure/repositories/challenge-for-correction-repository.js
@@ -1,6 +1,6 @@
-import * as solutionAdapter from '../../../shared/infrastructure/adapters/solution-adapter.js';
 import * as challengeRepository from '../../../shared/infrastructure/repositories/challenge-repository.js';
 import { ChallengeForCorrection } from '../../domain/models/ChallengeForCorrection.js';
+import * as solutionAdapter from '../adapters/solution-adapter.js';
 
 export async function get(challengeId) {
   const challenge = await challengeRepository.get_proxy(challengeId);
