+++ api/tests/evaluation/unit/domain/usecases/save-and-correct-answer-for-campaign_test.js
@@ -86,7 +86,7 @@ describe('Unit | Evaluation | Domain | Use Cases | save-and-correct-answer-for-c
     answer.result = undefined;
     answer.resultDetails = undefined;
     correctAnswerValue = '1';
-    solutionAlgo = domainBuilder.buildSolution({ id: answer.challengeId, value: correctAnswerValue });
+    solutionAlgo = domainBuilder.evaluation.buildSolution({ id: answer.challengeId, value: correctAnswerValue });
     challenge = domainBuilder.evaluation.buildChallengeForCorrection({
       id: answer.challengeId,
       solutionAlgo,
