+++ api/tests/tooling/domain-builder/factory/evaluation/index.js
@@ -1,9 +1,11 @@
 import { buildChallengeForCorrection } from './build-challenge-for-correction.js';
 import { buildChallengeToPlay } from './build-challenge-to-play.js';
 import { buildSmartRandomChallenge } from './build-smart-random-challenge.js';
+import { buildSolution } from './build-solution.js';
 
 export const builders = {
   buildSmartRandomChallenge,
   buildChallengeToPlay,
   buildChallengeForCorrection,
+  buildSolution,
 };
